Consolidate game information into one screen
Fixes #158
