+++ app/javascript/src/components/Rondel.vue
@@ -1,6 +1,6 @@
 <template>
   <svg
-    class="rondel"
+    class="rondel self-center"
     width="450px"
     height="450px"
     viewBox="-20 -20 240 240"
